buggy turn system
A basic turn system is working, it is a bit buggy but it might not be a problem for gameplay or performance.

If there ever is any problem it could be the mix of signals that happens.

and I will try to fix this right now but i might not